refactor: split discovery-client-lite into Python package
Refactor the single-file script into a Python package at
lite/discovery_client_lite/ with focused modules:

  models.py   - Endpoint, CachedReferral, ConnectedController, ClusterState
  metrics.py  - SimpleHistogram, Metrics, MetricsHandler, Prometheus endpoint
  nvme.py     - NvmeCli wrapper, sysfs readers, disconnect_by_nqn()
  config.py   - Config parsing, referral cache, YAML/env loading
  daemon.py   - DiscoveryDaemon, AEN listener, sd_notify
  cli.py      - Argparse, command handlers, run_serve
  __main__.py - Entry point for python -m invocation

Adds wrapper script (lite/bin/discovery-client-lite), Makefile install
target, and lb.yaml component for build system integration. Tests
migrated to lite/tests/ (72 tests passing).

Issue: LBM1-43309
