+++ lite/tests/test_discovery_client_lite.py
@@ -1,6 +1,5 @@
 """Tests for discovery-client-lite."""
 
-import importlib.util
 import os
 import shutil
 import sys
@@ -8,68 +7,69 @@ import tempfile
 import unittest
 from unittest.mock import patch, MagicMock
 
-# Load the hyphenated module via importlib
-_SCRIPT = os.path.join(os.path.dirname(__file__), '..', 'discovery-client-lite.py')
-_spec = importlib.util.spec_from_file_location('dcl', _SCRIPT)
-dcl = importlib.util.module_from_spec(_spec)
-_spec.loader.exec_module(dcl)
+from discovery_client_lite.models import Endpoint
+from discovery_client_lite.metrics import SimpleHistogram, Metrics
+from discovery_client_lite.nvme import NvmeCli, get_host_id
+from discovery_client_lite.config import parse_config_line, parse_interval, extract_referrals, load_env_overrides
+from discovery_client_lite.cli import build_parser, cmd_add_hostnqn, cmd_remove_hostnqn, _split_host_port
+from discovery_client_lite.daemon import DiscoveryDaemon
 
 
 class TestParseConfigLine(unittest.TestCase):
 
     def test_basic_line(self):
-        ep = dcl.parse_config_line('-t tcp -a 10.0.0.1 -s 8009 -q hostnqn1 -n subnqn1')
+        ep = parse_config_line('-t tcp -a 10.0.0.1 -s 8009 -q hostnqn1 -n subnqn1')
         self.assertEqual(ep.traddr, '10.0.0.1')
         self.assertEqual(ep.port, '8009')
         self.assertEqual(ep.hostnqn, 'hostnqn1')
         self.assertEqual(ep.subnqn, 'subnqn1')
 
     def test_comment_line(self):
-        self.assertIsNone(dcl.parse_config_line('# comment'))
+        self.assertIsNone(parse_config_line('# comment'))
 
     def test_empty_line(self):
-        self.assertIsNone(dcl.parse_config_line(''))
+        self.assertIsNone(parse_config_line(''))
 
     def test_missing_traddr(self):
-        self.assertIsNone(dcl.parse_config_line('-t tcp -s 8009'))
+        self.assertIsNone(parse_config_line('-t tcp -s 8009'))
 
     def test_default_port(self):
-        ep = dcl.parse_config_line('-a 10.0.0.1')
+        ep = parse_config_line('-a 10.0.0.1')
         self.assertEqual(ep.port, '8009')
 
     def test_dhchap_flags(self):
-        ep = dcl.parse_config_line('-a 10.0.0.1 -S secret1 -C ctrlsecret1')
+        ep = parse_config_line('-a 10.0.0.1 -S secret1 -C ctrlsecret1')
         self.assertEqual(ep.secret, 'secret1')
         self.assertEqual(ep.ctrl_secret, 'ctrlsecret1')
 
     def test_hostid_flag(self):
-        ep = dcl.parse_config_line('-a 10.0.0.1 -I uuid-123')
+        ep = parse_config_line('-a 10.0.0.1 -I uuid-123')
         self.assertEqual(ep.hostid, 'uuid-123')
 
     def test_ctrl_loss_tmo(self):
-        ep = dcl.parse_config_line('-a 10.0.0.1 -l 600')
+        ep = parse_config_line('-a 10.0.0.1 -l 600')
         self.assertEqual(ep.ctrl_loss_tmo, 600)
 
 
 class TestParseInterval(unittest.TestCase):
 
     def test_int_value(self):
-        self.assertEqual(dcl.parse_interval(5), 5)
+        self.assertEqual(parse_interval(5), 5)
 
     def test_seconds_string(self):
-        self.assertEqual(dcl.parse_interval('5s'), 5)
+        self.assertEqual(parse_interval('5s'), 5)
 
     def test_minutes_string(self):
-        self.assertEqual(dcl.parse_interval('1m'), 60)
+        self.assertEqual(parse_interval('1m'), 60)
 
     def test_plain_string(self):
-        self.assertEqual(dcl.parse_interval('10'), 10)
+        self.assertEqual(parse_interval('10'), 10)
 
     def test_negative_clamped(self):
-        self.assertEqual(dcl.parse_interval(-1), 0)
+        self.assertEqual(parse_interval(-1), 0)
 
     def test_invalid(self):
-        self.assertEqual(dcl.parse_interval('abc'), 0)
+        self.assertEqual(parse_interval('abc'), 0)
 
 
 class TestExtractReferrals(unittest.TestCase):
@@ -85,7 +85,7 @@ class TestExtractReferrals(unittest.TestCase):
                 },
             ]
         }
-        refs = dcl.extract_referrals(output)
+        refs = extract_referrals(output)
         self.assertEqual(len(refs), 1)
         self.assertEqual(refs[0].traddr, '10.0.0.2')
 
@@ -100,29 +100,29 @@ class TestExtractReferrals(unittest.TestCase):
                 },
             ]
         }
-        refs = dcl.extract_referrals(output)
+        refs = extract_referrals(output)
         self.assertEqual(len(refs), 0)
 
     def test_empty_records(self):
-        self.assertEqual(dcl.extract_referrals({}), [])
+        self.assertEqual(extract_referrals({}), [])
 
 
 class TestSimpleHistogram(unittest.TestCase):
 
     def test_observe_increments_count(self):
-        h = dcl.SimpleHistogram('test_hist', 'A test histogram')
+        h = SimpleHistogram('test_hist', 'A test histogram')
         h.observe(0.5)
         h.observe(1.5)
         self.assertEqual(h.count, 2)
 
     def test_observe_tracks_sum(self):
-        h = dcl.SimpleHistogram('test_hist', 'A test histogram')
+        h = SimpleHistogram('test_hist', 'A test histogram')
         h.observe(0.1)
         h.observe(0.2)
         self.assertAlmostEqual(h.total, 0.3)
 
     def test_bucket_boundaries(self):
-        h = dcl.SimpleHistogram('test_hist', 'A test histogram')
+        h = SimpleHistogram('test_hist', 'A test histogram')
         h.observe(0.003)  # <= 0.005 bucket
         h.observe(0.5)    # <= 0.5 bucket
         h.observe(100.0)  # +Inf only
@@ -133,7 +133,7 @@ class TestSimpleHistogram(unittest.TestCase):
         self.assertIn('test_hist_count 3', rendered)
 
     def test_render_format(self):
-        h = dcl.SimpleHistogram('test_hist', 'A test histogram')
+        h = SimpleHistogram('test_hist', 'A test histogram')
         rendered = h.render()
         self.assertIn('# HELP test_hist A test histogram', rendered)
         self.assertIn('# TYPE test_hist histogram', rendered)
@@ -144,7 +144,7 @@ class TestSimpleHistogram(unittest.TestCase):
 class TestMetrics(unittest.TestCase):
 
     def setUp(self):
-        self.m = dcl.Metrics()
+        self.m = Metrics()
 
     def test_render_uses_discovery_prefix(self):
         rendered = self.m.render()
@@ -181,47 +181,47 @@ class TestMetrics(unittest.TestCase):
 class TestLoadEnvOverrides(unittest.TestCase):
 
     def test_empty_env(self):
-        result = dcl.load_env_overrides({})
+        result = load_env_overrides({})
         self.assertEqual(result, {})
 
     def test_clientconfigdir(self):
         env = {'DC_CLIENTCONFIGDIR': '/tmp/test-config'}
-        result = dcl.load_env_overrides(env)
+        result = load_env_overrides(env)
         self.assertEqual(result['clientConfigDir'], '/tmp/test-config')
 
     def test_nested_logging(self):
         env = {'DC_LOGGING_LEVEL': 'debug', 'DC_LOGGING_FILENAME': '/tmp/test.log'}
-        result = dcl.load_env_overrides(env)
+        result = load_env_overrides(env)
         self.assertEqual(result['logging.level'], 'debug')
         self.assertEqual(result['logging.filename'], '/tmp/test.log')
 
     def test_both_interval_names(self):
         env = {'DC_RECONNECTINTERVAL': '10s'}
-        result = dcl.load_env_overrides(env)
+        result = load_env_overrides(env)
         self.assertEqual(result['reconnectInterval'], '10s')
 
         env2 = {'DC_POLLINGINTERVAL': '15s'}
-        result2 = dcl.load_env_overrides(env2)
+        result2 = load_env_overrides(env2)
         self.assertEqual(result2['pollingInterval'], '15s')
 
     def test_unrelated_env_ignored(self):
         env = {'HOME': '/home/user', 'PATH': '/usr/bin', 'DC_KATO': '20'}
-        result = dcl.load_env_overrides(env)
+        result = load_env_overrides(env)
         self.assertEqual(len(result), 1)
         self.assertEqual(result['kato'], '20')
 
     def test_autodetect_nested(self):
         env = {'DC_AUTODETECTENTRIES_ENABLED': 'false'}
-        result = dcl.load_env_overrides(env)
+        result = load_env_overrides(env)
         self.assertEqual(result['autoDetectEntries.enabled'], 'false')
 
 
 class TestNvmeCliDiscover(unittest.TestCase):
 
-    @patch.object(dcl.NvmeCli, '_run_json')
+    @patch.object(NvmeCli, '_run_json')
     def test_discover_builds_correct_args(self, mock_run):
         mock_run.return_value = {'records': []}
-        dcl.NvmeCli.discover(
+        NvmeCli.discover(
             traddr='10.0.0.1', trsvcid='8009', hostnqn='nqn.host1',
             transport='tcp',
         )
@@ -234,10 +234,10 @@ class TestNvmeCliDiscover(unittest.TestCase):
         self.assertIn('-q', args)
         self.assertIn('nqn.host1', args)
 
-    @patch.object(dcl.NvmeCli, '_run_json')
+    @patch.object(NvmeCli, '_run_json')
     def test_discover_persistent_flag(self, mock_run):
         mock_run.return_value = {}
-        dcl.NvmeCli.discover(
+        NvmeCli.discover(
             traddr='10.0.0.1', trsvcid='8009', hostnqn='nqn.host1',
             persistent=True,
         )
@@ -247,10 +247,10 @@ class TestNvmeCliDiscover(unittest.TestCase):
 
 class TestNvmeCliConnect(unittest.TestCase):
 
-    @patch.object(dcl.NvmeCli, '_run_json')
+    @patch.object(NvmeCli, '_run_json')
     def test_connect_builds_correct_args(self, mock_run):
         mock_run.return_value = {'cntlid': 1}
-        dcl.NvmeCli.connect(
+        NvmeCli.connect(
             traddr='10.0.0.1', trsvcid='4420', nqn='nqn.subsys1',
             transport='tcp',
         )
@@ -260,10 +260,10 @@ class TestNvmeCliConnect(unittest.TestCase):
         self.assertIn('-n', args)
         self.assertIn('nqn.subsys1', args)
 
-    @patch.object(dcl.NvmeCli, '_run_json')
+    @patch.object(NvmeCli, '_run_json')
     def test_connect_with_dhchap(self, mock_run):
         mock_run.return_value = {'cntlid': 1}
-        dcl.NvmeCli.connect(
+        NvmeCli.connect(
             traddr='10.0.0.1', trsvcid='4420', nqn='nqn.subsys1',
             dhchap_secret='secret1', dhchap_ctrl_secret='ctrl1',
         )
@@ -277,25 +277,25 @@ class TestNvmeCliConnect(unittest.TestCase):
 class TestNvmeCliMethods(unittest.TestCase):
 
     def test_disconnect_all_exists(self):
-        self.assertTrue(hasattr(dcl.NvmeCli, 'disconnect_all'))
-        self.assertTrue(callable(dcl.NvmeCli.disconnect_all))
+        self.assertTrue(hasattr(NvmeCli, 'disconnect_all'))
+        self.assertTrue(callable(NvmeCli.disconnect_all))
 
     def test_list_controllers_exists(self):
-        self.assertTrue(hasattr(dcl.NvmeCli, 'list_controllers'))
-        self.assertTrue(callable(dcl.NvmeCli.list_controllers))
+        self.assertTrue(hasattr(NvmeCli, 'list_controllers'))
+        self.assertTrue(callable(NvmeCli.list_controllers))
 
 
 class TestParseIntervalHours(unittest.TestCase):
 
     def test_hours_string(self):
-        self.assertEqual(dcl.parse_interval('96h'), 96 * 3600)
+        self.assertEqual(parse_interval('96h'), 96 * 3600)
 
     def test_hours_to_days(self):
-        hours = dcl.parse_interval('96h')
+        hours = parse_interval('96h')
         self.assertEqual(hours // 3600, 96)
 
     def test_48h(self):
-        self.assertEqual(dcl.parse_interval('48h'), 48 * 3600)
+        self.assertEqual(parse_interval('48h'), 48 * 3600)
 
 
 class TestGetHostId(unittest.TestCase):
@@ -304,12 +304,12 @@ class TestGetHostId(unittest.TestCase):
         with tempfile.NamedTemporaryFile(mode='w', suffix='.hostid', delete=False) as f:
             f.write('custom-uuid-123\n')
             f.flush()
-            result = dcl.get_host_id(f.name)
+            result = get_host_id(f.name)
             self.assertEqual(result, 'custom-uuid-123')
             os.unlink(f.name)
 
     def test_missing_file(self):
-        result = dcl.get_host_id('/nonexistent/path/hostid')
+        result = get_host_id('/nonexistent/path/hostid')
         self.assertEqual(result, '')
 
 
@@ -318,7 +318,7 @@ class TestArgparse(unittest.TestCase):
 
     def _parse(self, args_str):
         """Helper: build the parser and parse args."""
-        return dcl.build_parser().parse_args(args_str.split())
+        return build_parser().parse_args(args_str.split())
 
     def test_discover_flags(self):
         args = self._parse('discover -a 10.0.0.1 -q nqn.host1')
@@ -407,7 +407,7 @@ class TestAddHostnqn(unittest.TestCase):
         args.nqn = 'nqn.subsys1'
         args.transport = 'tcp'
 
-        rc = dcl.cmd_add_hostnqn(args, self.config_dir)
+        rc = cmd_add_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 0)
 
         filepath = os.path.join(self.config_dir, 'test-cluster')
@@ -427,7 +427,7 @@ class TestAddHostnqn(unittest.TestCase):
         args.transport = 'tcp'
         args.hostid = ''
 
-        rc = dcl.cmd_add_hostnqn(args, self.config_dir)
+        rc = cmd_add_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 1)
 
     def test_includes_hostid_when_provided(self):
@@ -439,7 +439,7 @@ class TestAddHostnqn(unittest.TestCase):
         args.nqn = 'nqn.subsys1'
         args.transport = 'tcp'
 
-        dcl.cmd_add_hostnqn(args, self.config_dir)
+        cmd_add_hostnqn(args, self.config_dir)
         content = open(os.path.join(self.config_dir, 'test-cluster')).read()
         self.assertIn('-I uuid-abc-123', content)
 
@@ -452,7 +452,7 @@ class TestAddHostnqn(unittest.TestCase):
         args.nqn = 'nqn.subsys1'
         args.transport = 'tcp'
 
-        dcl.cmd_add_hostnqn(args, self.config_dir)
+        cmd_add_hostnqn(args, self.config_dir)
         content = open(os.path.join(self.config_dir, 'test-cluster')).read()
         self.assertIn('-s 8009', content)
 
@@ -472,14 +472,14 @@ class TestRemoveHostnqn(unittest.TestCase):
 
         args = MagicMock()
         args.name = 'test-cluster'
-        rc = dcl.cmd_remove_hostnqn(args, self.config_dir)
+        rc = cmd_remove_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 0)
         self.assertFalse(os.path.exists(filepath))
 
     def test_missing_file_silent(self):
         args = MagicMock()
         args.name = 'nonexistent'
-        rc = dcl.cmd_remove_hostnqn(args, self.config_dir)
+        rc = cmd_remove_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 0)
 
 
@@ -488,20 +488,20 @@ class TestEndpointShuffling(unittest.TestCase):
     def test_poll_cycle_shuffles_endpoints(self):
         """Verify that endpoints are shuffled before failover iteration."""
         import inspect
-        source = inspect.getsource(dcl.DiscoveryDaemon.poll_cycle)
+        source = inspect.getsource(DiscoveryDaemon.poll_cycle)
         self.assertIn('random.shuffle', source)
 
 
 class TestSplitHostPort(unittest.TestCase):
 
     def test_ip_with_port(self):
-        self.assertEqual(dcl._split_host_port('10.0.0.1:8009'), ('10.0.0.1', '8009'))
+        self.assertEqual(_split_host_port('10.0.0.1:8009'), ('10.0.0.1', '8009'))
 
     def test_ip_without_port(self):
-        self.assertEqual(dcl._split_host_port('10.0.0.1'), ('10.0.0.1', '8009'))
+        self.assertEqual(_split_host_port('10.0.0.1'), ('10.0.0.1', '8009'))
 
     def test_ip_with_custom_port(self):
-        self.assertEqual(dcl._split_host_port('10.0.0.1:4420'), ('10.0.0.1', '4420'))
+        self.assertEqual(_split_host_port('10.0.0.1:4420'), ('10.0.0.1', '4420'))
 
 
 class TestAddHostnqnCommaSplit(unittest.TestCase):
@@ -522,7 +522,7 @@ class TestAddHostnqnCommaSplit(unittest.TestCase):
         args.nqn = 'nqn.subsys1'
         args.transport = 'tcp'
 
-        rc = dcl.cmd_add_hostnqn(args, self.config_dir)
+        rc = cmd_add_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 0)
 
         content = open(os.path.join(self.config_dir, 'test-cluster')).read()
@@ -530,7 +530,7 @@ class TestAddHostnqnCommaSplit(unittest.TestCase):
         self.assertEqual(len(lines), 3, f"Expected 3 lines, got {len(lines)}: {lines}")
         # Each line must have a clean IP in -a, not IP:PORT
         for line in lines:
-            ep = dcl.parse_config_line(line)
+            ep = parse_config_line(line)
             self.assertNotIn(':', ep.traddr, f"traddr should not contain port: {ep.traddr}")
             self.assertEqual(ep.port, '8009')
 
@@ -544,7 +544,7 @@ class TestAddHostnqnCommaSplit(unittest.TestCase):
         args.nqn = 'nqn.subsys1'
         args.transport = 'tcp'
 
-        rc = dcl.cmd_add_hostnqn(args, self.config_dir)
+        rc = cmd_add_hostnqn(args, self.config_dir)
         self.assertEqual(rc, 0)
 
         content = open(os.path.join(self.config_dir, 'test-cluster')).read()
@@ -556,18 +556,18 @@ class TestParseConfigLineWithPort(unittest.TestCase):
     """Defense-in-depth: parse_config_line handles IP:PORT in -a field."""
 
     def test_addr_with_embedded_port_no_s_flag(self):
-        ep = dcl.parse_config_line('-t tcp -a 10.0.0.1:8009 -q nqn.host1')
+        ep = parse_config_line('-t tcp -a 10.0.0.1:8009 -q nqn.host1')
         self.assertEqual(ep.traddr, '10.0.0.1')
         self.assertEqual(ep.port, '8009')
 
     def test_addr_with_embedded_port_and_s_flag(self):
         """Explicit -s takes precedence over embedded port."""
-        ep = dcl.parse_config_line('-t tcp -a 10.0.0.1:9999 -s 4420 -q nqn.host1')
+        ep = parse_config_line('-t tcp -a 10.0.0.1:9999 -s 4420 -q nqn.host1')
         # -s was given explicitly, so traddr keeps the full value
         self.assertEqual(ep.port, '4420')
 
     def test_addr_without_port(self):
-        ep = dcl.parse_config_line('-t tcp -a 10.0.0.1 -q nqn.host1')
+        ep = parse_config_line('-t tcp -a 10.0.0.1 -q nqn.host1')
         self.assertEqual(ep.traddr, '10.0.0.1')
         self.assertEqual(ep.port, '8009')
